Throw overflow exception when interval seconds are outside of range [0, 59] [databricks] (#9192)
* added GpuIntervalUtils shim for 350

* added a test

* singing off


* renamed trait

* added comment for method

---------
